requirements.txt
Django==3.2.14
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dj-rest-auth==2.2.4
django-cors-headers==3.13.0
bcrypt==3.2.0
PyMySQL==1.0.2
mysqlclient==2.1.1
python-dotenv==0.20.0
drf-yasg==1.20.0
djangorestframework-simplejwt==5.2.0
dj-rest-auth==2.2.4
django-redis==5.2.0
django-background-tasks==1.2.5
gunicorn==20.1.0
cryptography==37.0.4